+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## HEAD
+
+- Start tmux on ssh login.
+
 ## 0.1.0
 
 - Initial release.